+++ IO/PPM_Image.cpp
@@ -2,7 +2,7 @@
 
 PPM_Image::PPM_Image() { }
 
-PPM_Image::PPM_Image(unsigned int imageWidth, unsigned int imageHeight)
+PPM_Image::PPM_Image(uint16_t imageWidth, uint16_t imageHeight)
 //											  Multiply by 3, rgb values for each pixel.
 	: width(imageWidth), height(imageHeight), data(new uint8_t[imageWidth * imageHeight * 3])
 {
